create new functions: resetAttempts:resetAttempts resets the number of attempts for each salt in the given attempts map to the maximum number of attempts and updates the resetTime to be one day later if the current time is after the resetTime.
decrementAttempts: decrementAttempts decrements the number of attempts associated with the given salt in the provided map salt_with_attempt. If the salt is not found in the map, an error is returned. If the number of attempts reaches zero, an error is also returned. Otherwise, the attempts counter is decremented by one and the function returns nil.
Complete rate limit function.